fix review: moved SearchImmersionFiltersKey in back
