+++ back/src/domains/establishment/ports/SearchMadeRepository.ts
@@ -1,4 +1,31 @@
+import type {
+  ExtractAddedOrMissingKeys,
+  GetOffersFlatQueryParams,
+} from "shared";
 import type { SearchMadeEntity } from "../entities/SearchMadeEntity";
+
+const searchFilters = [
+  "fitForDisabledWorkers",
+  "searchableBy",
+  "nafCodes",
+  "remoteWorkModes",
+  "sirets",
+  "appellationCodes",
+  "locationIds",
+  "showOnlyAvailableOffers",
+] as const;
+
+type SearchFilterKey = (typeof searchFilters)[number];
+
+export type SearchImmersionFiltersKey =
+  | keyof Pick<GetOffersFlatQueryParams, SearchFilterKey>
+  | "geoParams"
+  | "romeCodes";
+
+export type ExtractAddedOrMissingSearchFiltersKeys<
+  T extends object,
+  I extends SearchImmersionFiltersKey = never,
+> = ExtractAddedOrMissingKeys<SearchImmersionFiltersKey, T, I>;
 
 export interface SearchMadeRepository {
   insertSearchMade: (searchMadeEntity: SearchMadeEntity) => Promise<void>;
